added static cast to avoid warnings...

+++ src/algorithm/171/cpp/excelnumber.hxx
@@ -7,7 +7,7 @@ public:
   int titleToNumber(std::string s) {
     int sum{};
 
-    for (int down{s.size() - 1}, up{}, coef26{1}; up < s.size(); (sum += (static_cast<int>(s[down]) - 64) * coef26), ++up, --down, coef26 *= 26);
+    for (int down{static_cast<int>(s.size()) - 1}, up{}, coef26{1}; up < s.size(); (sum += (static_cast<int>(s[down]) - 64) * coef26), ++up, --down, coef26 *= 26);
 
     return sum;
   }
